requirements.txt
requests
fastapi
uvicorn
python-dotenv
beautifulsoup4
google-generativeai
openai